Added deploy-base.py
This is a tool to help developers who want to work on Ames Stereo
Pipeline. After the user has produced a BaseSystem tarball, they can
use deploy base to install in spot on their hard drive and then fix
all the links. It also provides example config.options for VW and ASP
to get them started.