set_rpath: If not using relative install name, use absolute path
This is OSX specific, I don't think Linux really has any concept like
this.

Now the install name is always set in the event absolute paths are
used for the name. It will kick on no matter what, even if the path
isn't in the search range. This is important for the framework
libraries.