Add JAMA and TNT dependencies
Even though these are header only libraries. I'm not using that
specialization. My reason being is that they should go in the
noinstall directory. Noinstall directory was originally used to
identify libraries and applications that were shipped in ISIS. Now
that we plan to build ISIS ... that folder doesn't really serve any
demand.