Partially revert: 75e6821: Breaks libtool files
The regular expression in 75e6821 does nothing to fix the libtool
files. I've revert the code back to what works. Also the other control
files that 75e6821 adds to the mix, do not have any paths in them to
be corrected. I left them in the mix because correcting their
permissions is possibly something that needs to happen.