Fix compilation errors on MacOS 10.13
This fixes some compilation bugs on QT. An odd one is the removal of
error handling calls to labels. QT seems to have done this themselves on
the main branch.

This also cleans up the builds of QT. We were previously building both
the release and debug versions of QT which takes a lot of CPU time. I
changed it to only build release. I also disable more unused QT modules.

Finally this fixes a compilation error in liblas where clang does allow
mixing comparisons with integers and pointers.